Complain and skip for non-default input places.
That's weird because I would have thought it would be failing already
because the app is being invoked without params (or setting the
base_dir). But.... maybe locally (dev and test) it's not blowing away
the directory from previous tests in the "no lingering files" step, but
on Travis it is?

Anyway, there should not be lingering files in the default place that
cause it to wrongly pass the non-default tests. Matters once this new
skip is no longer a skip.