Nasty instrumentation of both new tshoot.py and thumbnails.py.
I should instead use some sort of sensible logging. Learn how to log in
behave. But for now, just trying to get some answers of whether the
files are getting moved on Travis's servers, and if not, how to tell.
Eventually should assert if they're not in the right place!